feat: Generic Consumers with retries
Breaking: Consumers are now generic
New: Reworks the "recv loop" to allow for retries.
If retries are configured, on handling error (not msg parsing) it will
Ack the message and spawn a goroutine to sleep and notify on the retry
channel.
This channel, on recv, will call the handler and keep track of the retry
count.