v2 (#6)
Revamp to the entire library to provide a much simpler public interface both for the connector and consumers/publishers.

Consumers are not typed anymore and the implementation function has a context.Context.
This means, implementations need to perform the type conversion themselves.

Producers/Publishers do not need to be 'implemented' anymore. A Publish method is now available directly in the connector that allows anyone to publish to the desired Exchange + Topic.